# doc/user/swagger.yaml
openapi: 3.0.0
info:
  title: Backend Pretest AI API
  description: API Documentation for Backend Pretest AI project
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
tags:
  - name: auth
    description: Authentication and Registration
  - name: user
    description: User Profile Management
paths:
  /auth/register:
    post:
      tags:
        - auth
      summary: Register a new user
      description: Create a new user account and send OTP to email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
        '400':
          description: Bad Request
  /auth/verify-otp:
    post:
      tags:
        - auth
      summary: Verify OTP for registration
      description: Verify the OTP sent to user email during registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyOTPRequest'
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '404':
          description: Not Found
  /auth/resend-otp:
    post:
      tags:
        - auth
      summary: Resend OTP
      description: Resend verification OTP to user email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResendOTPRequest'
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '404':
          description: Not Found
  /auth/login:
    post:
      tags:
        - auth
      summary: Login user
      description: Authenticate user and return JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Unauthorized
  /user/me:
    get:
      tags:
        - user
      summary: Get current user profile
      description: Get the profile data of the currently logged-in user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Unauthorized
        '404':
          description: Not Found
  /user/email/request-update:
    post:
      tags:
        - user
      summary: Request email update
      description: Send OTP to new email for verification
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEmailRequest'
      responses:
        '200':
          description: OK
  /user/email/verify-update:
    post:
      tags:
        - user
      summary: Verify email update
      description: Verify OTP and update user email
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyUpdateEmailRequest'
      responses:
        '200':
          description: OK

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
      required:
        - name
        - email
        - password
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    VerifyOTPRequest:
      type: object
      properties:
        email:
          type: string
        otp:
          type: string
      required:
        - email
        - otp
    UpdateEmailRequest:
      type: object
      properties:
        new_email:
          type: string
      required:
        - new_email
    VerifyUpdateEmailRequest:
      type: object
      properties:
        new_email:
          type: string
        otp:
          type: string
      required:
        - new_email
        - otp
    APIResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
        error:
          type: string
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/UserResponse'
    UserResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
        is_verified:
          type: boolean